solve comments from review
