+++ Sources/RSocketCore/Extensions/Coder/Encoder/DataEncoder.swift
@@ -1,3 +1,4 @@
+ODataEncoder.swift
 /*
  * Copyright 2015-present the original author or authors.
  *
@@ -37,7 +38,7 @@ extension DataDecoderProtocol {
 
 extension DataDecoderProtocol {
     @inlinable
-    func decode(from data: Foundation.Data) throws -> Data {
+    internal func decode(from data: Foundation.Data) throws -> Data {
         var buffer = ByteBuffer(data: data)
         let data = try self.decode(from: &buffer)
         guard buffer.readableBytes == 0 else {
