+++ Sources/RSocketCore/Extensions/Coder/Encoder/MetadataEncoder.swift
@@ -25,7 +25,7 @@ public protocol MetadataDecoder: CompositeMetadataDecoder {
 
 extension MetadataDecoder {
     @inlinable
-    func decode(from data: Data) throws -> Metadata {
+    public func decode(from data: Data) throws -> Metadata {
         var buffer = ByteBuffer(data: data)
         let metadata = try self.decode(from: &buffer)
         guard buffer.readableBytes == 0 else {
