+++ Sources/RSocketCore/Extensions/Coder/ErasableMetadata.swift
@@ -30,174 +30,3 @@ extension Optional: ErasableMetadata {
 extension Array: ErasableMetadata {
     public static var erasedValue: Array<Element> { [] }
 }
-
-extension Decoders {
-    public struct EraseMetadata<Decoder>: DecoderProtocol where 
-    Decoder: DecoderProtocol
-    {
-        @usableFromInline
-        internal var decoder: Decoder
-        
-        @usableFromInline
-        internal init(decoder: Decoder) {
-            self.decoder = decoder
-        }
-        
-        @inlinable
-        mutating public func decode(
-            _ payload: Payload, 
-            encoding: ConnectionEncoding
-        ) throws -> (Void, Decoder.Data) {
-            let (_, data) = try decoder.decode(payload, encoding: encoding)
-            return ((), data)
-        }
-    }
-}
-
-extension DecoderProtocol {
-    @inlinable
-    public func eraseMetadata() -> Decoders.EraseMetadata<Self> {
-        .init(decoder: self)
-    }
-}
-
-
-extension Decoders {
-    public struct PreserveMetadata<Decoder>: DecoderProtocol where 
-    Decoder: DecoderProtocol
-    {
-        @usableFromInline
-        internal var decoder: Decoder
-        
-        @usableFromInline
-        internal init(decoder: Decoder) {
-            self.decoder = decoder
-        }
-        
-        @inlinable
-        mutating public func decode(
-            _ payload: Payload, 
-            encoding: ConnectionEncoding
-        ) throws -> (Void, (Decoder.Metadata, Decoder.Data)) {
-            return ((), try decoder.decode(payload, encoding: encoding))
-        }
-    }
-}
-
-extension DecoderProtocol {
-    @inlinable
-    public func preserveMetadata() -> Decoders.PreserveMetadata<Self> {
-        .init(decoder: self)
-    }
-}
-
-
-extension Encoders {
-    public struct EraseMetadata<Encoder>: EncoderProtocol where 
-    Encoder: EncoderProtocol,
-    Encoder.Metadata: ErasableMetadata
-    {
-        @usableFromInline
-        internal var encoder: Encoder
-        
-        @usableFromInline
-        internal init(encoder: Encoder) {
-            self.encoder = encoder
-        }
-        
-        @inlinable
-        public mutating func encode(
-            metadata: Void, 
-            data: Encoder.Data, 
-            encoding: ConnectionEncoding
-        ) throws -> Payload {
-            try encoder.encode(
-                metadata: Encoder.Metadata.erasedValue, 
-                data: data, 
-                encoding: encoding
-            )
-        }
-    }
-}
-
-extension EncoderProtocol {
-    @inlinable
-    public func eraseMetadata() -> Encoders.EraseMetadata<Self> where Metadata: ErasableMetadata {
-        .init(encoder: self)
-    }
-}
-
-
-extension Encoders {
-    public struct PreserveMetadata<Encoder>: EncoderProtocol where 
-    Encoder: EncoderProtocol
-    {
-        @usableFromInline
-        internal var encoder: Encoder
-        
-        @usableFromInline
-        internal init(encoder: Encoder) {
-            self.encoder = encoder
-        }
-        
-        @inlinable
-        public mutating func encode(
-            metadata: Void, 
-            data: (Encoder.Metadata, Encoder.Data), 
-            encoding: ConnectionEncoding
-        ) throws -> Payload {
-            let (metadata, data) = data
-            return try encoder.encode(
-                metadata: metadata, 
-                data: data, 
-                encoding: encoding
-            )
-        }
-    }
-}
-
-extension EncoderProtocol {
-    @inlinable
-    public func preserveMetadata() -> Encoders.PreserveMetadata<Self> {
-        .init(encoder: self)
-    }
-}
-
-
-extension Encoders {
-    public struct SetMetadata<Encoder>: EncoderProtocol where 
-    Encoder: EncoderProtocol
-    {
-        @usableFromInline
-        internal var encoder: Encoder
-        
-        @usableFromInline
-        internal let metadata: Encoder.Metadata
-        
-        @usableFromInline
-        internal init(encoder: Encoder, metadata: Encoder.Metadata) {
-            self.encoder = encoder
-            self.metadata = metadata
-        }
-        
-        @inlinable
-        public mutating func encode(
-            metadata: Void, 
-            data: Encoder.Data, 
-            encoding: ConnectionEncoding
-        ) throws -> Payload {
-            try encoder.encode(
-                metadata: self.metadata, 
-                data: data, 
-                encoding: encoding
-            )
-        }
-    }
-}
-
-extension EncoderProtocol {
-    @inlinable
-    public func setMetadata(_ metadata: Metadata) -> Encoders.SetMetadata<Self> {
-        .init(encoder: self, metadata: metadata)
-    }
-}
